feat:실행 클래스와 콘솔 클래스 분리
