+++ src/main/java/Application.java
@@ -1,0 +1,7 @@
+public class Application {
+    public static void main(String[] args) {
+        Console console = new Console();
+        console.start();
+    }
+}
+
